+++ packages/python-packages/api-stub-generator/apistub/nodes/_docstring_parser.py
@@ -100,7 +100,11 @@ class DocstringParser:
             self.pos_args[arg.argname] = arg
         elif keyword == "keyword":
             # show kwarg is optional by setting default to "..."
+            # also wrap the type in Optional[] so it aligns with
+            # optionals identified in type hints.
             arg.default = "..."
+            if arg.argtype and not arg.argtype.startswith("Optional["):
+                arg.argtype = f"Optional[{arg.argtype}]"
             self.kw_args[arg.argname] = arg
         else:
             logging.error(f"Unexpected keyword: {keyword}")