+++ packages/python-packages/api-stub-generator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release History
 
+## Version 0.2.8 (Unreleased)
+Kwargs that were previously displayed as "type = ..." will now
+  be displayed as "Optional[type] = ..." to align with syntax
+  from Python type hints.
+Fixed issue where variadic arguments appeared a singular argument.
+Fixes issue where default values were no longer displayed.
+
 ## Version 0.2.7 (Unreleased)
 Updated version to regenerate all reviews using Python 3.9